New fix and features
prepare.sh - update php settings; create database for zabbix until created
prepare.sercice - now wanted by mysql server
insidefakeroot - create zabbix-frontend config; index.php as default page

+++ overlay/board/beaglebone/inside-fakeroot.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -x
 
 #########################################################
 #
@@ -6,20 +7,12 @@
 #
 #########################################################
 
-# disable systemd-networkd && systemd-resolved
-# TODO
-
 # rename default config 
-mv ${TARGET_DIR}/etc/systemd/network/dhcp.network ${TARGET_DIR/}/etc/systemd/network/beaglebone.network
+if [ -f  ${TARGET_DIR}/etc/systemd/network/dhcp.network ]; then
+    mv ${TARGET_DIR}/etc/systemd/network/dhcp.network ${TARGET_DIR/}/etc/systemd/network/beaglebone.network
+fi
 
-#########################################################
-#
-# make fstab
-#
-#########################################################
-fstabFile="${TARGET_DIR}/etc/fstab"
-echo "" > $fstabFile
-echo "LABEL=$dataImageFsLabel /data $dataImageFsType  defaults   0 1" >> $fstabFile
+
 
 #########################################################
 #
@@ -39,11 +32,10 @@ done
 
 #########################################################
 #
-# add apache php support
+# apache
 #
 #########################################################
-
-set -x
+# Add php
 grep -q 'SetHandler application/x-httpd-php' ${TARGET_DIR}/etc/apache2/httpd.conf
 php_apache_status=$?
 echo @$php_apache_status@
@@ -54,6 +46,12 @@ if [ $php_apache_status -gt 0 ]; then
 </FilesMatch>
 EOF
 fi
+# Set index.php as default
+rm ${TARGET_DIR}/usr/htdocs/index.html
+
+sed -i -e 's/DirectoryIndex index.html/DirectoryIndex index.php/g' ${TARGET_DIR}/etc/apache2/httpd.conf
+
+
 
 
 #########################################################
@@ -65,6 +63,42 @@ fi
 # Set dbPassword for zabbix server
 zabbix_password_status=grep -q 'DBPassword=zabbix' ${TARGET_DIR}/etc/zabbix_server.conf || \
     sed -i -e '/DBPassword=/a\' -e 'DBPassword=zabbix' ${TARGET_DIR}/etc/zabbix_server.conf
+
+#########################################################
+#
+# move  zabbix frontend to html  dir
+#
+#########################################################
+
+if [ -d ${TARGET_DIR}/usr/zabbix/php-frontend ]; then
+    mv ${TARGET_DIR}/usr/zabbix/php-frontend/* ${TARGET_DIR}/usr/htdocs/
+    rm -rf ${TARGET_DIR}/usr/zabbix/php-frontend
+fi
+
+cat > ${TARGET_DIR}/usr/htdocs/conf/zabbix.conf.php <<EOF
+<?php
+// Zabbix GUI configuration file.
+global $DB;
+
+\$DB['TYPE']     = 'MYSQL';
+\$DB['SERVER']   = 'localhost';
+\$DB['PORT']     = '0';
+\$DB['DATABASE'] = 'zabbix';
+\$DB['USER']     = 'zabbix';
+\$DB['PASSWORD'] = 'zabbix';
+
+// Schema name. Used for IBM DB2 and PostgreSQL.
+\$DB['SCHEMA'] = '';
+
+\$ZBX_SERVER      = 'localhost';
+\$ZBX_SERVER_PORT = '10051';
+\$ZBX_SERVER_NAME = '';
+
+\$IMAGE_FORMAT_DEFAULT = IMAGE_FORMAT_PNG;
+EOF
+
+
+
 
 #########################################################
 #
@@ -88,14 +122,20 @@ fi
 if [ -d $dataImageMountDir ]; then
     rm -rf $dataImageMountDir
 fi
-set -x
 dd if=/dev/zero of=$dataImageFile bs=16M count=1
 mkfs.$dataImageFsType -L $dataImageFsLabel $dataImageFile
 mkdir -p $dataImageMountDir
 sudo mount -t $dataImageFsType -o loop $dataImageFile $dataImageMountDir
 sudo cp ${TARGET_DIR/}/etc/systemd/network/beaglebone.network  $dataImageMountDir/
 sudo cp ${TARGET_DIR}/etc/zabbix_server.conf $dataImageMountDir/
-sudo cp ${TARGET_DIR}/etc/zabbix_agent.conf $dataImageMountDir/
+sudo cp ${TARGET_DIR}/etc/zabbix_agentd.conf $dataImageMountDir/
 sudo umount $dataImageMountDir
 
-mkdir ${TARGET_DIR}/data/
+#########################################################
+#
+# make fstab
+#
+#########################################################
+fstabFile="${TARGET_DIR}/etc/fstab"
+echo "" > $fstabFile
+echo "LABEL=$dataImageFsLabel /data $dataImageFsType  defaults   0 1" >> $fstabFile
